python_backend/requirements.txt
torch>=2.0.0
transformers>=4.35.0
pillow>=10.0.0
flask>=3.0.0
flask-cors>=4.0.0
modelscope>=1.9.0
accelerate>=0.24.0

